+++ aws_auth.tf
@@ -1,12 +1,12 @@
 resource "local_file" "config_map_aws_auth" {
-  count    = var.write_aws_auth_config ? 1 : 0
-  content  = data.template_file.config_map_aws_auth.rendered
+  count    = var.create_cluster && var.write_aws_auth_config ? 1 : 0
+  content  = data.template_file.config_map_aws_auth.0.rendered
   filename = "${var.config_output_path}config-map-aws-auth_${var.cluster_name}.yaml"
 }
 
 resource "null_resource" "update_config_map_aws_auth" {
-  count      = var.manage_aws_auth ? 1 : 0
-  depends_on = [aws_eks_cluster.this]
+  count      = var.create_cluster && var.manage_aws_auth ? 1 : 0
+  depends_on = [aws_eks_cluster.this.0]
 
   provisioner "local-exec" {
     working_dir = path.module
@@ -26,21 +26,22 @@ EOS
   }
 
   triggers = {
-    kube_config_map_rendered = data.template_file.kubeconfig.rendered
-    config_map_rendered      = data.template_file.config_map_aws_auth.rendered
-    endpoint                 = aws_eks_cluster.this.endpoint
+    kube_config_map_rendered = data.template_file.kubeconfig.0.rendered
+    config_map_rendered      = data.template_file.config_map_aws_auth.0.rendered
+    endpoint                 = aws_eks_cluster.this.0.endpoint
   }
 }
 
 data "aws_caller_identity" "current" {
+  count = var.create_cluster ? 1 : 0
 }
 
 data "template_file" "launch_template_worker_role_arns" {
-  count    = local.worker_group_launch_template_count
+  count    = var.create_cluster ? local.worker_group_launch_template_count : 0
   template = file("${path.module}/templates/worker-role.tpl")
 
   vars = {
-    worker_role_arn = "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/${element(
+    worker_role_arn = "arn:aws:iam::${data.aws_caller_identity.current.0.account_id}:role/${element(
       coalescelist(
         aws_iam_instance_profile.workers_launch_template.*.role,
         data.aws_iam_instance_profile.custom_worker_group_launch_template_iam_instance_profile.*.role_name,
@@ -56,11 +57,11 @@ data "template_file" "launch_template_worker_role_arns" {
 }
 
 data "template_file" "worker_role_arns" {
-  count    = local.worker_group_count
+  count    = var.create_cluster ? local.worker_group_count : 0
   template = file("${path.module}/templates/worker-role.tpl")
 
   vars = {
-    worker_role_arn = "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/${element(
+    worker_role_arn = "arn:aws:iam::${data.aws_caller_identity.current.0.account_id}:role/${element(
       coalescelist(
         aws_iam_instance_profile.workers.*.role,
         data.aws_iam_instance_profile.custom_worker_group_iam_instance_profile.*.role_name,
@@ -77,6 +78,7 @@ data "template_file" "worker_role_arns" {
 }
 
 data "template_file" "config_map_aws_auth" {
+  count    = var.create_cluster ? 1 : 0
   template = file("${path.module}/templates/config-map-aws-auth.yaml.tpl")
 
   vars = {