+++ cluster.tf
@@ -1,5 +1,5 @@
 resource "aws_cloudwatch_log_group" "this" {
-  count             = length(var.cluster_enabled_log_types) > 0 ? 1 : 0
+  count             = var.create_cluster && length(var.cluster_enabled_log_types) > 0 ? 1 : 0
   name              = "/aws/eks/${var.cluster_name}/cluster"
   retention_in_days = var.cluster_log_retention_in_days
   kms_key_id        = var.cluster_log_kms_key_id
@@ -7,6 +7,7 @@ resource "aws_cloudwatch_log_group" "this" {
 }
 
 resource "aws_eks_cluster" "this" {
+  count                     = var.create_cluster ? 1 : 0
   name                      = var.cluster_name
   enabled_cluster_log_types = var.cluster_enabled_log_types
   role_arn                  = local.cluster_iam_role_arn
@@ -33,7 +34,7 @@ resource "aws_eks_cluster" "this" {
 }
 
 resource "aws_security_group" "cluster" {
-  count       = var.cluster_create_security_group ? 1 : 0
+  count       = var.create_cluster && var.cluster_create_security_group ? 1 : 0
   name_prefix = var.cluster_name
   description = "EKS cluster security group."
   vpc_id      = var.vpc_id
@@ -46,7 +47,7 @@ resource "aws_security_group" "cluster" {
 }
 
 resource "aws_security_group_rule" "cluster_egress_internet" {
-  count             = var.cluster_create_security_group ? 1 : 0
+  count             = var.create_cluster && var.cluster_create_security_group ? 1 : 0
   description       = "Allow cluster egress access to the Internet."
   protocol          = "-1"
   security_group_id = local.cluster_security_group_id
@@ -57,7 +58,7 @@ resource "aws_security_group_rule" "cluster_egress_internet" {
 }
 
 resource "aws_security_group_rule" "cluster_https_worker_ingress" {
-  count                    = var.cluster_create_security_group ? 1 : 0
+  count                    = var.create_cluster && var.cluster_create_security_group ? 1 : 0
   description              = "Allow pods to communicate with the EKS cluster API."
   protocol                 = "tcp"
   security_group_id        = local.cluster_security_group_id
@@ -68,9 +69,9 @@ resource "aws_security_group_rule" "cluster_https_worker_ingress" {
 }
 
 resource "aws_iam_role" "cluster" {
-  count                 = var.manage_cluster_iam_resources ? 1 : 0
+  count                 = var.create_cluster && var.manage_cluster_iam_resources ? 1 : 0
   name_prefix           = var.cluster_name
-  assume_role_policy    = data.aws_iam_policy_document.cluster_assume_role_policy.json
+  assume_role_policy    = data.aws_iam_policy_document.cluster_assume_role_policy.0.json
   permissions_boundary  = var.permissions_boundary
   path                  = var.iam_path
   force_detach_policies = true
@@ -78,13 +79,13 @@ resource "aws_iam_role" "cluster" {
 }
 
 resource "aws_iam_role_policy_attachment" "cluster_AmazonEKSClusterPolicy" {
-  count      = var.manage_cluster_iam_resources ? 1 : 0
+  count      = var.create_cluster && var.manage_cluster_iam_resources ? 1 : 0
   policy_arn = "arn:aws:iam::aws:policy/AmazonEKSClusterPolicy"
   role       = local.cluster_iam_role_name
 }
 
 resource "aws_iam_role_policy_attachment" "cluster_AmazonEKSServicePolicy" {
-  count      = var.manage_cluster_iam_resources ? 1 : 0
+  count      = var.create_cluster && var.manage_cluster_iam_resources ? 1 : 0
   policy_arn = "arn:aws:iam::aws:policy/AmazonEKSServicePolicy"
   role       = local.cluster_iam_role_name
 }